+++ InAppStorySdk/inappstorysdk/src/main/java/com/inappstory/sdk/stories/ui/list/StoryFavoriteListItem.java
@@ -207,7 +207,7 @@ public class StoryFavoriteListItem extends BaseStoryListItem {
         if (InAppStoryService.isNull()) return;
         Downloader.downloadFileBackground(url, InAppStoryService.getInstance().getFastCache(), new FileLoadProgressCallback() {
             @Override
-            public void onProgress(int loadedSize, int totalSize) {
+            public void onProgress(long loadedSize, long totalSize) {
 
             }
